add_validatenagano_cake

+++ app/controllers/public/orders_controller.rb
@@ -23,7 +23,7 @@ class Public::OrdersController < ApplicationController
     elsif params[:order][:address_option] == "2"
       @order.postal_code = params[:order][:postal_code]
       @order.address = params[:order][:address]
-      @order.name = params[:order][:name]
+      @order.name = params[:order][:name]       #if分を用いて、空欄の時にエラー文をフラッシュメッセージで表示させる
 
     else
       render 'new'
@@ -33,22 +33,25 @@ class Public::OrdersController < ApplicationController
 
   end
 
-  def comfilm
+  def create
     @order = Order.new(order_params)
     @order.customer_id = current_customer.id
-    @order.save!
+    if @order.save
 
-    current_customer.cart_items.each do |cart_item|
-      @order_detail = OrderDetail.new
-      @order_detail.order_id = @order.id #商品詳細に注文IDを紐付け
-      @order_detail.item_id = cart_item.item_id#商品idを注文詳細idに代入
-      @order_detail.amount = cart_item.amount#商品の個数を注文詳細に代入
-      @order_detail.price = cart_item.item.with_tax_price#消費税込みの値段
-      @order_detail.save
+      current_customer.cart_items.each do |cart_item|
+        @order_detail = OrderDetail.new
+        @order_detail.order_id = @order.id #商品詳細に注文IDを紐付け
+        @order_detail.item_id = cart_item.item_id#商品idを注文詳細idに代入
+        @order_detail.amount = cart_item.amount#商品の個数を注文詳細に代入
+        @order_detail.price = cart_item.item.with_tax_price#消費税込みの値段
+        @order_detail.save
+      end
+
+      current_customer.cart_items.destroy_all
+      redirect_to orders_complete_path
+    else
+      render 'new'
     end
-
-    current_customer.cart_items.destroy_all
-    redirect_to public_orders_complete_path
 
   end
